backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pandas==2.2.0
openpyxl==3.1.2
apify-client==1.6.3 
requests==2.31.0
pytz==2024.1
python-dotenv==1.0.1 